# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Сладкая API
  version: 0.1.0
servers:
  - url: https://api.example.invalid
paths:
  /healthz:
    get:
      operationId: health
      responses:
        "200":
          description: Service is healthy
  /v1/device/provision:
    post:
      operationId: provisionDevice
      description: One-time mobile provisioning endpoint. Production exposure and apiOrigin must use HTTPS.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceProvisionInput"
      responses:
        "201":
          description: Device bearer credential returned exactly once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceProvisioned"
        "400":
          description: JSON or activation proof format is malformed
        "401":
          description: Activation is unknown, expired, already consumed, or does not match the expected device proof
        "500":
          description: Credential generation or atomic persistence failed; no credential is returned
  /v1/device/measurements:
    post:
      operationId: ingestMeasurement
      security:
        - deviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeasurementInput"
      responses:
        "202":
          description: Measurement accepted idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "400":
          description: Invalid measurement
        "401":
          description: Invalid device token
        "409":
          description: Device binding mismatch or conflicting event identity
  /v1/family/session:
    post:
      operationId: issueFamilySession
      description: Server-side BFF exchange. The provisioned family access token must never be returned to browser JavaScript.
      security:
        - familyAccess: []
      parameters:
        - $ref: "#/components/parameters/FamilyOrigin"
      responses:
        "201":
          description: Secure family_session cookie plus a session-bound CSRF token
          headers:
            Set-Cookie:
              description: family_session with HttpOnly, Secure, SameSite=Strict, Path=/ and bounded Max-Age attributes
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilySessionIssued"
        "401":
          description: Provisioned family access is missing, invalid or expired
        "403":
          description: Origin is not in the configured exact allowlist
  /v1/patients/{patientId}/snapshot:
    get:
      operationId: getPatientSnapshot
      security:
        - familySession: []
      parameters:
        - $ref: "#/components/parameters/PatientId"
      responses:
        "200":
          description: Latest patient state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientSnapshot"
        "401":
          description: Family session cookie is missing or invalid
        "404":
          description: Patient or measurement not found
  /v1/patients/{patientId}/measurements:
    get:
      operationId: listMeasurements
      security:
        - familySession: []
      parameters:
        - $ref: "#/components/parameters/PatientId"
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Measurements in ascending sensor-time order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Measurement"
        "401":
          description: Family session cookie is missing or invalid
  /v1/alerts/{alertId}/acknowledge:
    post:
      operationId: acknowledgeAlert
      security:
        - familySession: []
      parameters:
        - in: path
          name: alertId
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/FamilyOrigin"
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
            pattern: "^[A-Za-z0-9_-]{43}$"
      responses:
        "204":
          description: Alert acknowledged idempotently
        "400":
          description: Alert identifier is not a UUID
        "401":
          description: Family session is required
        "403":
          description: Origin or CSRF token is invalid
        "404":
          description: Alert does not belong to the configured patient
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
    familyAccess:
      type: http
      scheme: bearer
      description: Provisioned high-entropy family access material accepted only by the server-side session exchange.
    familySession:
      type: apiKey
      in: cookie
      name: family_session
      description: Server-issued session cookie; must use HttpOnly, Secure and SameSite=Strict attributes. Authorization Bearer is not accepted.
  parameters:
    FamilyOrigin:
      in: header
      name: Origin
      required: true
      schema:
        type: string
        format: uri
    PatientId:
      in: path
      name: patientId
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    FamilySessionIssued:
      type: object
      additionalProperties: false
      required: [csrfToken, expiresAt]
      properties:
        csrfToken:
          type: string
          pattern: "^[A-Za-z0-9_-]{43}$"
          description: Session-bound token for the BFF to send as X-CSRF-Token on family mutations.
        expiresAt:
          type: string
          format: date-time
    DeviceProvisionInput:
      type: object
      additionalProperties: false
      required: [activationCode, deviceId, deviceNonce]
      properties:
        activationCode:
          type: string
          pattern: "^SLK1-(?:[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{4}-){7}[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{4}$"
          description: Exact 160-bit one-time activation code created by the admin CLI.
        deviceId:
          type: string
          format: uuid
        deviceNonce:
          type: string
          pattern: "^[A-Za-z0-9_-]{43}$"
          description: Stable canonical base64url encoding of 32 random bytes generated by this app installation.
    DeviceProvisioned:
      type: object
      additionalProperties: false
      required: [deviceToken, apiOrigin, deviceId, patientId, backendBindingId, credentialId, credentialRevision]
      properties:
        deviceToken:
          type: string
          pattern: "^[A-Za-z0-9_-]{43}$"
          description: Bearer credential returned only by the successful one-time response.
        apiOrigin:
          type: string
          format: uri
          description: Exact configured API origin; HTTPS is mandatory in production.
        deviceId:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        backendBindingId:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
        credentialId:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
        credentialRevision:
          type: integer
          minimum: 1
          maximum: 9007199254740991
    MeasurementInput:
      type: object
      unevaluatedProperties: false
      required:
        - deviceId
        - backendBindingId
        - credentialId
        - credentialRevision
      properties:
        deviceId:
          type: string
          format: uuid
        backendBindingId:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
        credentialId:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
        credentialRevision:
          type: integer
          minimum: 1
          maximum: 9007199254740991
      allOf:
        - $ref: "#/components/schemas/MeasurementFields"
    MeasurementFields:
      type: object
      required:
        - eventId
        - sensorId
        - sensorFamily
        - sensorTime
        - phoneTime
        - glucoseMgDl
        - trendMgDlPerMinute
        - quality
        - sequence
      properties:
        eventId:
          oneOf:
            - type: string
              format: uuid
            - type: string
              pattern: "^[0-9a-f]{64}$"
        sensorId:
          type: string
          minLength: 1
          maxLength: 128
        sensorFamily:
          type: string
          enum: [sibionics_gs1, sibionics_gs1sb, sibionics_gs3]
        sensorTime:
          type: string
          format: date-time
        phoneTime:
          type: string
          format: date-time
        glucoseMgDl:
          type: integer
          minimum: 20
          maximum: 600
        trendMgDlPerMinute:
          type: number
          minimum: -20
          maximum: 20
        quality:
          type: string
          enum: [valid, warming_up, degraded]
        sequence:
          type: integer
          minimum: 0
          maximum: 9007199254740991
    Measurement:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/MeasurementFields"
        - type: object
          required: [patientId, receivedAt]
          properties:
            patientId:
              type: string
              format: uuid
            receivedAt:
              type: string
              format: date-time
    IngestResult:
      type: object
      additionalProperties: false
      required: [accepted]
      properties:
        accepted:
          type: boolean
          const: true
    PatientSnapshot:
      type: object
      required: [patientId, freshness, latest, openAlerts]
      properties:
        patientId:
          type: string
          format: uuid
        freshness:
          type: string
          enum: [fresh, stale, missing]
        latest:
          oneOf:
            - $ref: "#/components/schemas/Measurement"
            - type: "null"
        openAlerts:
          type: array
          items:
            type: object
            required: [id, kind, openedAt]
            properties:
              id:
                type: string
                format: uuid
              kind:
                type: string
                enum: [low, high, rapid_fall, rapid_rise, signal_loss]
              openedAt:
                type: string
                format: date-time
              acknowledgedAt:
                oneOf:
                  - type: string
                    format: date-time
                  - type: "null"